Move music global action handling to an own component
Due to requiring components that are added at an OsuGame-level
